backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Vector Store
chromadb==0.6.3
onnxruntime==1.20.1

# Document Processing
python-docx==1.1.2
pypdf>=5.1.0

# LLM Providers
groq==0.13.1
openai==1.58.1

# Utilities
python-jose==3.3.0
httpx==0.28.1
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
